backend/requirements_windows.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements_windows.txt requirements.in
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via web3
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.1.0
    # via aiohttp
bitarray==3.0.0
    # via eth-account
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
ckzg==2.0.1
    # via eth-account
click==8.1.8
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   uvicorn
cytoolz==1.0.1
    # via eth-utils
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via -r requirements.in
eth-abi==5.2.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.5
    # via web3
eth-hash[pycryptodome]==0.7.1
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.6.1
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.2.0
    # via eth-account
eth-typing==5.1.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.2.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
fastapi==0.115.8
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hexbytes==1.3.0
    # via
    #   eth-account
    #   eth-rlp
    #   web3
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.3
    # via pandas
packaging==24.2
    # via
    #   build
    #   pytest
pandas==2.2.3
    # via -r requirements.in
parsimonious==0.10.0
    # via eth-abi
pip-tools==7.4.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pycryptodome==3.21.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.10.6
    # via
    #   eth-account
    #   fastapi
    #   web3
pydantic-core==2.27.2
    # via pydantic
pymongo==4.11.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.25.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.1
    # via pandas
pyunormalize==16.0.0
    # via web3
pywin32==308
    # via web3
pyyaml==6.0.2
    # via uvicorn
regex==2024.11.6
    # via parsimonious
requests==2.32.3
    # via web3
rlp==4.1.0
    # via
    #   eth-account
    #   eth-rlp
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
toolz==1.0.0
    # via cytoolz
types-requests==2.32.0.20241016
    # via web3
typing-extensions==4.12.2
    # via
    #   anyio
    #   eth-typing
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   web3
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
uvicorn[standard]==0.34.0
    # via -r requirements.in
watchfiles==1.0.4
    # via uvicorn
web3==7.8.0
    # via -r requirements.in
websockets==13.1
    # via
    #   uvicorn
    #   web3
wheel==0.45.1
    # via pip-tools
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
